pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "steam-python"
version = "0.0.1"
description = "A fast and lightweight Python interface for interacting with Steam."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "aiohttp",
    "pycryptodome",
]

[project.optional-dependencies]
dev = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "grpcio-tools",
    "pylint",
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
"Homepage" = "https://python.steam.style"
"Bug Tracker" = "https://github.com/Steam-Style/steam/issues"

[tool.pylint.main]
ignore = ["CVS", "protobufs", "enums"]

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "unused-argument",
]